+++ login-workflow/example/src/App.tsx
@@ -4,14 +4,20 @@ import {
     AuthNavigationContainer,
     AuthUIContextProvider,
     useSecurityActions,
+    // RegistrationData,
 } from '@pxblue/react-auth-workflow';
 import { ProjectAuthUIActions } from './actions/AuthUIActions';
 import { ProjectRegistrationUIActions } from './actions/RegistrationUIActions';
 import { ExampleHome } from './screens/ExampleHome';
 // import { CustomDetailsScreen, CustomDetailsScreenTwo } from './components/CustomDetailsScreen';
 import { routes } from './navigation/Routing';
-
 import productLogo from './assets/images/eaton_stacked_logo.png';
+// import { Route } from 'react-router-dom';
+// import { ExamplePreAuth } from './screens/ExamplePreAuth';
+// import {
+//     CustomAccountAlreadyExistsScreen,
+//     CustomRegistrationSuccessScreen,
+// } from './screens/ExampleRegistrationSuccess';
 
 // Imports for internationalization
 import i18n from 'i18next';
@@ -67,6 +73,7 @@ export const AuthUIConfiguration: React.FC = (props) => {
                 </div>
             }
 
+            // loginType={'username'}
             // Uncomment this line to see how to add custom form fields to the registration screens
             // customAccountDetails={[
             //     { component: CustomDetailsScreen },
@@ -76,6 +83,10 @@ export const AuthUIConfiguration: React.FC = (props) => {
             //         instructions: 'Enter your employment information below.',
             //     },
             // ]}
+            // registrationSuccessScreen={(registrationData: RegistrationData): JSX.Element => (
+            //     <CustomRegistrationSuccessScreen registrationData={registrationData} />
+            // )}
+            // accountAlreadyExistsScreen={(): JSX.Element => <CustomAccountAlreadyExistsScreen />}
         >
             {props.children}
         </AuthUIContextProvider>
@@ -85,7 +96,10 @@ export const AuthUIConfiguration: React.FC = (props) => {
 export const App: React.FC = () => (
     <SecurityContextProvider>
         <AuthUIConfiguration>
-            <AuthNavigationContainer routeConfig={routes}>
+            <AuthNavigationContainer
+                routeConfig={routes}
+                // extraRoutes={[<Route exact path={'/pre-auth'} component={ExamplePreAuth} key={'pre-auth'} />]}
+            >
                 <ExampleHome />
             </AuthNavigationContainer>
         </AuthUIConfiguration>