+++ login-workflow/src/screens/Login.tsx
@@ -6,8 +6,8 @@ import {
     useAccountUIActions,
     useAccountUIState,
     useInjectedUIContext,
-    EMAIL_REGEX,
     AccountActions,
+    EMAIL_REGEX,
 } from '@pxblue/react-auth-shared';
 import { Link } from 'react-router-dom';
 import {
@@ -31,18 +31,24 @@ import clsx from 'clsx';
 
 const useStyles = makeStyles((theme: Theme) =>
     createStyles({
-        formFields: {
-            marginBottom: theme.spacing(3) + 22, // height of error message
+        emailFormField: {
+            marginBottom: theme.spacing(5),
             '&$hasError': {
                 marginBottom: theme.spacing(3),
             },
+            [theme.breakpoints.down('xs')]: {
+                marginBottom: theme.spacing(4),
+            },
+        },
+        passwordFormField: {
+            marginBottom: theme.spacing(3),
         },
         buttonRow: {
             marginBottom: theme.spacing(5),
             flexWrap: 'nowrap',
             [theme.breakpoints.down('xs')]: {
                 flexWrap: 'wrap',
-                flexDirection: 'column-reverse',
+                flexDirection: 'column',
                 justifyContent: 'center',
             },
         },
@@ -64,6 +70,9 @@ const useStyles = makeStyles((theme: Theme) =>
             padding: `${theme.spacing(4)}px ${theme.spacing(8)}px`,
             display: 'flex',
             flexDirection: 'column',
+            [theme.breakpoints.down('xs')]: {
+                padding: `${theme.spacing(4)}px ${theme.spacing(4)}px`,
+            },
         },
         link: {
             fontWeight: 600,
@@ -82,6 +91,11 @@ const useStyles = makeStyles((theme: Theme) =>
             width: 60,
             height: 60,
             color: theme.palette.text.secondary,
+        },
+        rememberMeCheckbox: {
+            [theme.breakpoints.down('xs')]: {
+                marginRight: 0,
+            },
         },
         hasError: {},
         productLogo: {
@@ -121,6 +135,7 @@ export const Login: React.FC = () => {
                 <img className={classes.productLogo} src={projectImage || stackedEatonLogo} alt="logo" />
             </div>
         ),
+        loginType = 'email',
         loginActions,
     } = useInjectedUIContext();
 
@@ -148,6 +163,10 @@ export const Login: React.FC = () => {
 
     const hasTransitError = authUIState.login.transitErrorMessage !== null;
     const transitErrorMessage = authUIState.login.transitErrorMessage ?? t('pxb:MESSAGES.REQUEST_ERROR');
+
+    const isInvalidCredentials =
+        transitErrorMessage.replace('pxb:', '') === 'LOGIN.INCORRECT_CREDENTIALS' ||
+        transitErrorMessage.replace('pxb:', '') === 'LOGIN.INVALID_CREDENTIALS';
 
     useEffect(
         () => {
@@ -260,31 +279,31 @@ export const Login: React.FC = () => {
                     {debugLinks}
 
                     <TextField
-                        label={t('pxb:LABELS.EMAIL')}
+                        label={loginType === 'username' ? t('pxb:LABELS.USERNAME') : t('pxb:LABELS.EMAIL')}
                         id="email"
-                        name="email"
-                        type="email"
-                        className={clsx(classes.formFields, { [classes.hasError]: hasTransitError })}
+                        name={loginType === 'username' ? 'username' : 'email'}
+                        type={loginType === 'username' ? 'text' : 'email'}
+                        className={clsx(classes.emailFormField, { [classes.hasError]: hasTransitError })}
                         value={emailInput}
                         onChange={(evt: ChangeEvent<HTMLInputElement>): void => setEmailInput(evt.target.value)}
                         onKeyPress={(e): void => {
                             if (e.key === 'Enter' && passwordField.current) passwordField.current.focus();
                         }}
                         variant="filled"
-                        error={hasTransitError}
-                        helperText={hasTransitError ? t('pxb:LOGIN.INCORRECT_CREDENTIALS') : ''}
+                        error={isInvalidCredentials}
+                        helperText={isInvalidCredentials ? t('pxb:LOGIN.INCORRECT_CREDENTIALS') : ''}
                     />
                     <SecureTextField
                         inputRef={passwordField}
                         id="password"
                         name="password"
                         label={t('pxb:LABELS.PASSWORD')}
-                        className={clsx(classes.formFields, { [classes.hasError]: hasTransitError })}
+                        className={clsx(classes.passwordFormField, { [classes.hasError]: hasTransitError })}
                         value={passwordInput}
                         onChange={(evt: ChangeEvent<HTMLInputElement>): void => setPasswordInput(evt.target.value)}
                         variant="filled"
-                        error={hasTransitError}
-                        helperText={hasTransitError ? t('pxb:LOGIN.INCORRECT_CREDENTIALS') : ''}
+                        error={isInvalidCredentials}
+                        helperText={isInvalidCredentials ? t('pxb:LOGIN.INCORRECT_CREDENTIALS') : ''}
                     />
 
                     <Grid
@@ -296,6 +315,7 @@ export const Login: React.FC = () => {
                     >
                         {showRememberMe && (
                             <FormControlLabel
+                                className={classes.rememberMeCheckbox}
                                 control={
                                     <Checkbox
                                         color="primary"
@@ -310,7 +330,11 @@ export const Login: React.FC = () => {
                             type="submit"
                             variant="contained"
                             disableElevation
-                            disabled={!EMAIL_REGEX.test(emailInput) || !passwordInput}
+                            disabled={
+                                loginType === 'username'
+                                    ? !emailInput || !passwordInput
+                                    : !EMAIL_REGEX.test(emailInput) || !passwordInput
+                            }
                             color="primary"
                             style={{ width: showRememberMe ? 150 : '100%' }}
                             onClick={loginTapped}